Move sqlite store integration tests out of module root
